some changes.

+++ mm-app-api/src/main/java/com/example/demo/controller/UserController.java
@@ -30,7 +30,7 @@ public class UserController {
 	}
 
 	@GetMapping("/{id}")
-	public ResponseEntity<User> getUserById(@PathVariable Long id) {
+	public ResponseEntity<User> getUserById(@PathVariable int id) {
 		Optional<User> user = userService.getUserById(id);
 		return user.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
 	}
@@ -41,7 +41,7 @@ public class UserController {
 	}
 
 	@PutMapping("/{id}")
-	public ResponseEntity<User> updateUser(@PathVariable Long id, @RequestBody User userDetails) {
+	public ResponseEntity<User> updateUser(@PathVariable int id, @RequestBody User userDetails) {
 		try {
 			User updatedUser = userService.updateUser(id, userDetails);
 			return ResponseEntity.ok(updatedUser);
@@ -51,7 +51,7 @@ public class UserController {
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+	public ResponseEntity<Void> deleteUser(@PathVariable int id) {
 		userService.deleteUser(id);
 		return ResponseEntity.noContent().build();
 	}
